+++ abzagent/core/__init__.py
@@ -29,6 +29,8 @@ try:
         tool_input_guardrail,
         tool_output_guardrail,
         GuardrailFunctionOutput,
+        GuardrailResult,
+        Guardrail,
         InputGuardrailTripwireTriggered,
         OutputGuardrailTripwireTriggered,
         ToolGuardrailTripwireTriggered,
@@ -36,6 +38,7 @@ try:
 except Exception:
     input_guardrail = output_guardrail = GuardrailFunctionOutput = None  # type: ignore
     tool_input_guardrail = tool_output_guardrail = None  # type: ignore
+    GuardrailResult = Guardrail = None  # type: ignore
     InputGuardrailTripwireTriggered = OutputGuardrailTripwireTriggered = None  # type: ignore
     ToolGuardrailTripwireTriggered = None  # type: ignore
 
@@ -54,7 +57,7 @@ __all__ = [
     # Optional
     "input_guardrail", "output_guardrail",
     "tool_input_guardrail", "tool_output_guardrail",
-    "GuardrailFunctionOutput",
+    "GuardrailFunctionOutput", "GuardrailResult", "Guardrail",
     "InputGuardrailTripwireTriggered", "OutputGuardrailTripwireTriggered",
     "ToolGuardrailTripwireTriggered",
 ]